requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
opencv-python>=4.5.3
numpy>=1.19.5
tensorflow>=2.8.0
pillow>=8.3.1
ultralytics>=8.0.0
python-jose>=3.3.0
python-dotenv>=0.19.0
